save and load of markers per user working

+++ markers table sql.sql
@@ -29,8 +29,8 @@ CREATE TABLE `markers` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `name` varchar(60) NOT NULL,
   `address` varchar(80) NOT NULL,
-  `lat` float(10,6) NOT NULL,
-  `lng` float(10,6) NOT NULL,
+  `lat` double(10,6) NOT NULL,
+  `lng` double(10,6) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
